GIFT EVERYONE EVERYTHING THEY COULD WANT WITH LORA (#366)

+++ include/f_core/messaging/c_latest_message_port.h
@@ -46,7 +46,7 @@ public:
     /**
      * See parent docs
      */
-    int AvailableSpace() override {
+    size_t AvailableSpace() override {
         return 1;
     }
